backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# AI and OpenAI
openai==1.3.7
tiktoken==0.5.2

# HTTP requests and API clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# HeyGen API integration
python-dotenv==1.0.0

# File handling and storage
python-multipart==0.0.6
aiofiles==23.2.1

# Background tasks and async
celery==5.3.4
redis==5.0.1

# Export and document generation
python-pptx==0.6.23
markdown==3.5.1
jinja2==3.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS and middleware
slowapi==0.1.9